Fix error logging and passing to client
Closes: https://github.com/metarhia/metacom/issues/251
