+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Remove `toString` in `receiveBody` to be compatible with ws
 - Fix parse broken JSON packets
 - Fix detecting ping packets (empty objects)
+- Fix error logging and passing to client
 
 ## [2.0.2][] - 2021-09-11
