+++ lib/channel.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const http = require('http');
 const metautil = require('metautil');
 
 const EMPTY_PACKET = Buffer.from('{}');
@@ -146,6 +147,19 @@ class Channel {
     this.send({ callback: callId, result });
     const record = `${this.ip}\t${interfaceName}/${methodName}`;
     application.console.log(record);
+  }
+
+  error(code, err = null, callId) {
+    const { req, ip, application } = this;
+    const { url, method } = req;
+    const status = http.STATUS_CODES[code];
+    const message = status || err.message || 'Unknown error';
+    const httpCode = status ? code : 500;
+    const reason = `${httpCode}\t${err ? err.stack : message}`;
+    application.console.error(`${ip}\t${method}\t${url}\t${reason}`);
+    const error = { message, code };
+    const packet = callId ? { callback: callId, error } : { error };
+    this.send(packet, httpCode);
   }
 
   async restoreSession() {
