+++ lib/http.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const http = require('http');
 const metautil = require('metautil');
 const { Channel } = require('./channel.js');
 
@@ -49,19 +48,6 @@ class HttpChannel extends Channel {
   send(obj, httpCode = 200) {
     const data = JSON.stringify(obj);
     this.write(data, httpCode, 'json');
-  }
-
-  error(code, err = null, callId) {
-    const { req, ip, application } = this;
-    const { url, method } = req;
-    const status = http.STATUS_CODES[code];
-    const reason = err ? err.stack : status;
-    application.console.error(`${ip}\t${method}\t${url}\t${code}\t${reason}`);
-    const message = status || 'Unknown error';
-    const httpCode = status ? code : 500;
-    const error = { message, code };
-    if (callId) this.send({ callback: callId, error });
-    else this.send({ error }, httpCode);
   }
 
   redirect(location) {
