+++ lib/ws.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const http = require('http');
 const { Channel } = require('./channel.js');
 
 class WsChannel extends Channel {
@@ -21,16 +20,6 @@ class WsChannel extends Channel {
 
   send(obj) {
     this.write(JSON.stringify(obj));
-  }
-
-  error(code, err = null, callId) {
-    const { req, ip, application } = this;
-    const { url, method } = req;
-    const status = http.STATUS_CODES[code];
-    const reason = err ? err.stack : status;
-    application.console.error(`${ip}\t${method}\t${url}\t${code}\t${reason}`);
-    const message = err ? err.message : code.toString();
-    this.send({ callback: callId, error: { message, code } });
   }
 }
